requirements.txt
# Basic utilities
requests
python-dotenv
pandas

# Data visualization
matplotlib

# Testing
pytest==7.4.0
pytest-cov==4.0.0
nbval

# Code formatting and linting
black==22.3.0
pylint==2.15.3
click==8.1.3 

# FastAPI and related utilities
fastapi==0.85.0
uvicorn[standard]==0.18.3
httpx

# Database connector
mysql-connector-python

# Load test
locust

